MCP/ECommerce/pyproject.toml
[project]
name = "ecommerce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.10.1",
    "mysql-connector-python>=9.3.0",
    "python-dotenv>=1.1.1",
]
